Fix data fetching error in used and new cars component
